requirements.txt
Flask==3.0.3
python-pptx==0.6.21
google-generativeai==0.7.2
